backend/requirements.txt
numpy==1.24.3
Flask==2.2.2
Werkzeug==2.2.2
pandas==2.2.0
python-dotenv
pillow
nltk
PyMuPDF
sentence_transformers
pinecone-client